[java-exceptions] Initial commit with base code structure
This commit contains the initial application structure for the
exception handling example. It contains a set of model classes and
a business logic class with a main() method, which will be the
entry point for running this example.

The model classes contain very basic logic and validation -- the
checks are performed and returned as simple boolean values, what
considerably limits the way an application can handle any logical
errors and fail cases. In the next commits this logic will be
enhanced and a better strategy will be implemented to work with
fail scenarios in the method execution flow.

Read the comments in the code to see which pitfalls are possible
in the current implementation, and check the next commits to see
how it is possible to improve this code.

NOTE: the current example will NOT RUN successfully. An exception
will be thrown at the last line, and is explained in the comment
block below it. We will fix it in the future commits.