[java-exceptions] Initialize lists to remove possible NPEs
In this commit, the credit cards list from the Account class is
initialized with an empty list from the Collections class to avoid
the NullPointerException observed earlier from happening during
application runtime.

This initially solves this problem, but as we need to add a new
card to the list later, it creates another issue due to its
UnsupportedOperationException, that is thrown because emptyList
represents an immutable collection.

We can learn through this example that an emptyList should not be
used for lists that may be possibly changed. To fix this problem,
we will manually create a list in the Account class constructor in
the next commit.

Some additional validations will be replaced to make use of the
available Java exceptions, so have a look at the diff for the next
commit to see what can be improved in our current checks.