[java-exceptions] Manually initialize list to avoid any exception
This commit initializes the credits card list in the Account class
manually with a new ArrayList to avoid any of the exceptions seen
before due to null values or unsupported operations. Now we can
add credit cards to the account without any issues.

However, the PIN code validation is not performed when setting a
new code. We will do that in the next commit, along with throwing
a programmatic exception if the desired new code does not match
the requirements.