[java-exceptions] Enhance checks and throw exceptions in error cases
This commit demonstrates how to throw an exception from inside the
application if the behavior is different than expected. Some methods
were enhanced with more detailed validations and now throw exceptions
if the arguments do not comply with the requirements. Some examples
are:
- PIN codes for credit cards should have six digits;
- Withdrawal or deposit amounts should not be negative values;
- An account can not be created prior to its customer.

The main ExceptionHandlingScenario was also improved with an example
of the classic strategy to work with exceptions in a Java program:
try/catch blocks, that allow the programmers to defined an expected
behavior that can throw an exception and what to do if it is indeed
thrown.

As we are now familiar of the basic approach to deal with potential
exceptions and also know how to throw them from our application code,
the next commit will introduce a customized exception to represent
that the customer does not have enough funds to perform a specific
operation.

The next commits will show more advanced techniques to work with
exceptions in Java, and also introduce some additional features.