[java-exceptions] Create a customized runtime exception
This commit introduces a new type of exception, created especially
for our application: the InsufficientFundsException. As a customer
can try to withdraw more money than his current credit limit, this
exception comes to the rescue and defines a very specific error
condition in the application.

Using our own exceptions allows us to better debug the application
as we will have a better context of what went wrong during its
execution -- it will only be thrown under a very specific condition
and that allows us to know what was wrong.

As it extends the java.lang.RuntimeException class, it will be an
unchecked exception. That means the Java compiler (and IDEs) will
not complain if we don't handle it explicitly in our application,
and the code will compile successfully - the problem will only be
evidenced when the code is executed.

In the current state, the code will not execute successfully. An
exception will be logged in the console in the last method call of
our main method, and this will show us that our custom exception
works as expected.