[java-exceptions] Convert InsufficientFundsException to checked
In this commit, our custom exception InsufficientFundsException
was converted from a runtime (also called unchecked) to a compile
time (also called checked) exception. As a checked exception, it
is recognized by the Java compiler and Java IDEs as a potential
exception thrown by the application.

This change makes the compiler aware of the need for exception
handling inside our code and, consequently, we (developers) are
require to handle the exception in our code. There are different
strategies for handling it:

- The method who explicitly throws the exception is required to
  update its signature with a 'throws' keyword to notify any
  potential caller that it can throw an exception at runtime and
  it must be handled by the caller.
- The callers of a method that can throw this exception have two
  possibilities for handling it:
  -- Wrap the method invocation in a try/catch block and handle
     it appropriately in the catch block;
  -- Update their signatures with the 'throws' keyword to inform
     its callers that it may throw an exception.

The second option will make the exception go up in the call stack
until a method handles it. If no one takes care of the exception,
the application will crash -- just as observed in the last commit.