[java-exceptions] Demonstrate 'finally' block in invoice generation
This commit enhances the CreditCard class to generate an invoice and
save it into a file when requested by the application. To implement
this feature, as the application needs to consume a system resource,
it was possible to introduce and demonstrate an usage example for
the 'finally' keyword. This keyword may be used to call operations
that should run independently of the exception being thrown in the
'try' block. In our application, we used it to do clean-up on the
system resources allocated to save the file.

In this commit we also discuss a little bit further about all the
possible execution flows for an application that uses the 'finally'
keyword in one of its exception handling blocks.