[java-exceptions] Catch multiple exception types in a single block
Here we introduce a slight change in our logic that generates an
invoice for a credit card. As IOException and NullPointerException
were handled with the exact same piece of code and required two
different 'catch' blocks to achieve that, we demonstrated how to
reduce it to a single block.

Java SE 7 introduced a new feature that reduces duplicated code and
avoids catching general exceptions instead of more specific ones.
By using a vertical bar (|), it is now possible to catch more than
one type of exception in the same catch block. This is what is
demonstrated in the changes submitted in this commit.