[java-exceptions] Use try-with-resources to write the invoice file
This commit inserts a structural change in CreditCard.createInvoice()
to demonstrate how the try-with-resources statement can be used when
handling system resources in a Java program. This functionality has
been added in Java 7 and provides a better way of managing resources
that should be closed within a try/catch statement.

Every object which implements the java.lang.AutoCloseable interface
(which included any object that implements java.io.Closeable) is
automatically closed by the JVM at the end of the statement. As this
interface defines the close() method, it is automatically called to
close the resource after the statement is executed.

To do that, we declare and instantiate the resource between brackets
when declaring the try block. Additionally, if the finally block is
only used to handle the closing of system resources, it may even be
entirely removed from the code, as this responsibility is transferred
to the JVM.