[java-exceptions] Update README with commit list
As the session demonstrating exception handling in Java is finished,
this commit aims on updating the README file to provide a list of
commits pushed into the repo for this series. By providing the users
a clean version of this list, I believe it will be easier for them
to find all modifications and steps on this tutorial.

When the repository grows with additional folders and tutorials,
there will be much more commits in the main commit list and things
will probably get harder to follow once multiple tutorial start to
get in the way of the desired commits.

Our main class, ExceptionHandlingScenario.java, was also modified
to provide a "happy ending" for John's story, and to make it clear
that the tutorial is now over.