[spring-mvc] Add essential Spring file structure
This comit introduces the essential files and configuration needed
for a web application based on the Spring Framework (including Spring
MVC) to work correctly.

A README file was also introduced to provide more details about some
importing and execution problems that may happen while trying to run
this code in Eclipse.

This application can be used as a quick start for any legacy Spring
application that needs just Spring MVC to work as intended. It has a
predefined controller and a dispatcher servlet establishing a very simple
navigation between two pages.